organize and group robot classes
