+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/teamcode/bypower/Decode2025RobotCode_TeleOp.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.teamcode;
+package org.firstinspires.ftc.teamcode.teamcode.bypower;
 
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
